refactor: fix package design to have layer structure
