+++ cmd/template/init.go
@@ -3,7 +3,7 @@ package template
 import (
 	"log"
 
-	"github.com/ArthurKC/scaffold/pkg/file"
+	atemplate "github.com/ArthurKC/scaffold/pkg/adapters/template"
 	"github.com/spf13/cobra"
 )
 
@@ -17,7 +17,7 @@ var initCmd = &cobra.Command{
 	Short: "Init directory as template.",
 	Long:  `Init directory as template. Create scaffold.yaml meta file.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		meta := file.NewMetaFile(args[0])
+		meta := atemplate.NewMetaFile(args[0])
 		if err := meta.Initialize(); err != nil {
 			log.Fatal(err)
 		}
