+++ pkg/adapters/generator/output.go
@@ -1,4 +1,4 @@
-package file
+package generator
 
 import (
 	"fmt"
